# openAPI/openapi.oas3.yml
openapi: 3.0.3
info:
    title: III-Exhibition works-HAI
    description: API specification for interactions between Frontend, Backend, and Cloud Storage.
    version: 1.0.0
servers:
    - url: https://frontend.example.com
      description: Frontend Server
paths:
    /generate:
        post:
            summary: Generate AI Drawing
            description: Post id in storage to BE.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                userId:
                                    type: string
                                    description: user id of the experience
                                    example: "hogehogehoge"
                            required:
                                - userId
            responses:
                "200":
                    description: Drawing saved successfully.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    url:
                                        type: string
                                        example: "http://hogehoge"
                                required:
                                    - url
                "400":
                    description: Invalid request - the request could not be understood due to malformed syntax or invalid data.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: "Invalid input data: user id is missing"
                default:
                    description: Unexpected error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        description: "A detailed error message."
    /init:
        get:
            summary: Get presigned urls
            description: Retrieve presigned URLs for both Human and AI drawings.
            responses:
                "200":
                    description: Presigned URLs for drawings retrieved successfully.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    result:
                                        type: object
                                        properties:
                                            id: 
                                                type: string
                                                description: "user id"
                                            urls:
                                                type: object
                                                properties:
                                                    humanDrawing:
                                                        type: string
                                                        description: "Presigned URL for human drawing upload."
                                                    topDrawing:
                                                        type: string
                                                    rightDrawing:
                                                        type: string
                                                    bottomDrawing:
                                                        type: string
                                                    leftDrawing:
                                                        type: string
                                                required:
                                                    - humanDrawing
                                        required:
                                            - id
                                            - urls
                                required:
                                    - result
                            example:
                                {
                                    "result":
                                        {
                                            "humanDrawing": "https://storage.googleapis.com/hai-human-ai-interaction.appspot.com/fuga.png?x-goog-signature=0fdc9fe30bfdb5d113b80c7c5795e6a2b8d5edeb9840cfd3ce6e81c51847a4b2c3958c3a249e42434a4d06ff811ae1c4de0ec71e2f478dd0852fe447a376f49ade17966138c70eaf879c6f563bc98516588be4ecee9b96038698e004ff511f1502dd8bc96d02c8d7b2d04ddea9659ec6f4aa5df912f0a21bd5dc19ce52b443bcc2dd414a627a9de9608dd3cb3761edd125e119834dc36d26bfe8a1986765e333862d402d61b858abe7dc8a46237198ccccba47ab3de98df41166c8dc1312e955a39a6a582122505edf01a3b5eb172437cd91e7445a4f4dbcb0fc35e62a1bda003c1e59bc043c0bf747a17301d31dced2266070c1c2315a31d74d15058d2d22b9&x-goog-algorithm=GOOG4-RSA-SHA256&x-goog-credential=firebase-adminsdk-n0huy%40hai-human-ai-interaction.iam.gserviceaccount.com%2F20240607%2Fasia-northeast1%2Fstorage%2Fgoog4_request&x-goog-date=20240607T034507Z&x-goog-expires=600&x-goog-signedheaders=host",
                                            "topDrawing": "https://example-cloud-storage.com/user_drawing.png?signature=...",
                                            "rightDrawing": "https://example-cloud-storage.com/user_drawing.png?signature=...",
                                            "bottomDrawing": "https://example-cloud-storage.com/user_drawing.png?signature=...",
                                            "leftDrawing": "https://example-cloud-storage.com/user_drawing.png?signature=...",
                                        },
                                }
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: "Invalid input data: user id is missing"
                "500":
                    description: Internal Server Error - the server encountered an unexpected condition that prevented it from fulfilling the request.
                default:
                    description: Unexpected error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        description: "A detailed error message."
    /view:
        get:
            summary: Drawing Viewer Page
            description: Viewer Page for human AI drawings. 
            responses:
                "200":
                    description: Drawings were successfully fetched.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    result:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                position:
                                                    type: object
                                                    properties:
                                                        x:
                                                            type: integer
                                                            description: target of x position
                                                            example: 0
                                                        y:
                                                            type: integer
                                                            description: target of y position
                                                            example: 0
                                                    required:
                                                        - x
                                                        - y
                                                url:
                                                    type: string
                                                    description: "Presigned URL for target position."
                                            required:
                                                - position
                                                - url
                                required:
                                    - result

                            example:
                                result: [
                                    {
                                        position: {
                                            x: 0,
                                            y: 0
                                        },
                                        url: "https://storage.googleapis.com/..."
                                    },
                                    {
                                        position: {
                                            x: 0,
                                            y: -1
                                        },
                                        url: "https://storage.googleapis.com/..."
                                    }
                                
                                ]
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: "Invalid input data"
                "404":
                    description: Not Found - No drawing found for the specified position.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        description: "A detailed error message."
                                example:
                                    error: "No drawing found for the specified position."
                "500":
                    description: Internal Server Error - The server encountered an unexpected condition that prevented it from fulfilling the request.
                default:
                    description: Unexpected error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        description: "A detailed error message."
    /rebase-database:
        get:
            summary: rebase database
            description: refresh position informations to clear data apparently
            responses:
                "200":
                    description: Drawings were successfully fetched.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                required:
                                    - success

                            example:
                                success: true

                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: "Invalid input data"
                "404":
                    description: Not Found - No drawing found for the specified position.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        description: "A detailed error message."
                                example:
                                    error: "No drawing found for the specified position."
                "500":
                    description: Internal Server Error - The server encountered an unexpected condition that prevented it from fulfilling the request.
                default:
                    description: Unexpected error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        description: "A detailed error message."

components:
    securitySchemes:
        ApiKeyAuth:
            type: apiKey
            in: header
            name: X-Api-Key
security:
    - ApiKeyAuth: []